test: merge small ingest test files into primaries
- test_source_ingest_json_list → test_source_ingest (JSON list edge cases)
- test_drive_ingest → test_drive (Drive ingestion + failure tracking)

Net: -2 files. All tests preserved.
